moved code to middleware, and some cleaning
