+++ backend/controllers/userController.js
@@ -9,10 +9,13 @@ const userController = {};
 
 // Function to register a new user, we must validate the information given by the frontend.
 userController.createUser = async (req, res) => {
+
     // Change the password given by frontend to the encrypted password
     req.body.password = await bcrypt.hash(req.body.password, 8);
+
     // Adds the uploaded photo filename (inside /public/userPhotos/) to request body before saving on DB
     req.body.photo = path.win32.basename(req.file.path);
+    
     // Verify the user type and use the respective schema
     if(req.body.role=="landlord"){
         const landlordDetail = new Landlord(req.body);
